Refactor: Slice Public API 적용 (index.ts) - 아직 meeting 엔티티 살아있음 추후 수정
