+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
 import '../src/app/globals.css';
-import { LayoutHeader } from '@/widgets';
+import { LayoutHeader } from '@/widgets/app-header';
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
