--- src/entities/index.ts
@@ -1,3 +1,0 @@
-// meeting
-export type { Meeting, MeetingStatus } from './meeting/model/types';
-export { mockMeetings } from './meeting/lib/mockMeetings';
