+++ src/entities/meeting/index.ts
@@ -1,0 +1,2 @@
+export type { Meeting, MeetingStatus } from './model/types';
+export { mockMeetings } from './lib/mockMeetings';
