+++ src/entities/member/index.ts
@@ -1,0 +1,1 @@
+// entity/member/index.ts
