+++ src/entities/note/index.ts
@@ -1,0 +1,1 @@
+// entity/note/index.ts
