+++ src/features/auth/signup/index.ts
@@ -1,0 +1,1 @@
+export { SignupForm } from './ui/SignupForm';
