--- src/features/index.ts
@@ -1,7 +1,0 @@
-// auth
-export { SignupForm } from './auth/signup/ui/SignupForm';
-export { LoginForm } from './auth/login/ui/LoginForm';
-
-// meeting
-export { CreateMeetingForm } from './meeting/create-room/ui/CreateMeetingForm';
-export { InviteModal } from './meeting/invite-modal/ui/InviteModal';
