+++ src/features/meeting/create-room/index.ts
@@ -1,0 +1,1 @@
+export { CreateMeetingForm } from './ui/CreateMeetingForm';
