+++ src/features/meeting/invite-modal/index.ts
@@ -1,0 +1,1 @@
+export { InviteModal } from './ui/InviteModal';
