--- src/shared/index.ts
@@ -1,23 +1,0 @@
-// ui
-export { Input, Button, Modal } from './ui';
-
-// lib
-export type { ValidationResult } from './lib';
-export {
-  validateEmail,
-  validatePassword,
-  validateLoginPassword,
-  validatePasswordConfirm,
-  validateNickname,
-  validateRoomPassword,
-  AUTH_STORAGE_KEY,
-  AUTH_CHANGE_EVENT,
-  markUserAuthenticated,
-  clearAuthState,
-  isUserAuthenticated,
-  copyToClipboard,
-  shareToKakao,
-  shareToSMS,
-  shareToEmail,
-} from './lib';
-
