--- src/shared/ui/index.ts
@@ -1,3 +1,0 @@
-export { Input } from './Input';
-export { Button } from './Button';
-export { Modal } from './Modal';
