+++ src/shared/ui/modal/index.ts
@@ -1,0 +1,1 @@
+export { Modal } from './Modal';
