+++ src/widgets/app-header/index.ts
@@ -1,0 +1,2 @@
+export { AppHeader } from './ui/AppHeader';
+export { LayoutHeader } from './ui/LayoutHeader';
