--- src/widgets/index.ts
@@ -1,6 +1,0 @@
-// app-header
-export { AppHeader } from './app-header/ui/AppHeader';
-export { LayoutHeader } from './app-header/ui/LayoutHeader';
-
-// site-footer
-export { SiteFooter } from './site-footer/ui/SiteFooter';
