+++ src/widgets/site-footer/index.ts
@@ -1,0 +1,1 @@
+export { SiteFooter } from './ui/SiteFooter';
